requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.0
faiss-cpu==1.8.0
numpy<2
pypdf==4.1.0
python-multipart==0.0.9
python-dotenv==1.0.1
pymongo>=4.6.0
